Deduplicate the shared pcap reader between benchmark.py and check_in_pyodide.py
An independent review of the previous commit's rationale for leaving
this duplication in place found it didn't hold up: both scripts'
"standalone" docstrings scope independence to netprotocols (the thing
each script verifies), the same usage CONTRIBUTING.md and
check_fixtures.py's own docstring use — not independence from each
other. Nothing forbids sharing a helper between them.

Extracted the near-identical read_pcap()/corpus_frames() pair into
scripts/_pcap.py: stdlib-only, no netprotocols import, so it preserves
the actual property the docstrings document. check_fixtures.py's own
read_pcap (materially different — it validates linktype and length)
is untouched, since it was never part of this duplication.

Verified against a real Pyodide runtime after the extraction: the
sys.path.insert(0, .../"scripts") + `from _pcap import corpus_frames`
import resolves correctly under Pyodide's mounted filesystem, and the
full 97-frame corpus still decodes unchanged.
