+++ CHANGELOG.md
@@ -60,10 +60,21 @@ methodology, and updated 1.1/1.2/1.6 figures.
   exception the way this project's own `ProtocolError` does, so a
   bare `except` is the correct interop boundary, not a swallowed bug.
   The flagged duplication between `benchmark.py` and
-  `check_in_pyodide.py`'s minimal pcap readers is left as-is — both
-  scripts' docstrings already document that each verification path
-  must stand on its own, independent of the others, which a shared
-  helper would undo for a cosmetic DRY gain.
+  `check_in_pyodide.py`'s minimal pcap readers — nearly byte-identical
+  `read_pcap`/`corpus_frames` functions — is now shared via a new
+  `scripts/_pcap.py`. An initial pass left the duplication in place,
+  reasoning both scripts' docstrings required each verification path
+  to stand fully independent of the others; re-reading them shows that
+  claim is narrower than stated — "standalone" there means not
+  depending on netprotocols (the thing each script verifies), the same
+  usage `CONTRIBUTING.md` and `check_fixtures.py`'s own docstring use,
+  never independence from sibling scripts. `_pcap.py` is itself
+  stdlib-only and does not import netprotocols, so it preserves the
+  property the docstrings actually document. Verified against a real
+  Pyodide runtime after the extraction: the corpus still decodes
+  unchanged, and `check_fixtures.py`'s own, materially different
+  `read_pcap` (it additionally validates linktype and length) is
+  untouched, since it was never part of this duplication.
 
 ### Documentation
 - **Corrected six README.md claims an independent audit against the
